ARMOA005-1487 Plano de carga bodegas, permitir decimales
Se realizaron las modificaciones en base de datos, backend y frontend para permitir el ingreso de decimales en las cantidades por bodega utilizando una coma como separador

+++ Molinos.Scato.Dominio/Dto/PlanoDeCargaBodegaDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int BodegaParcel { get; set; }
-        public int Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
         public MaterialPuertoDto MaterialPuerto { get; set; }
         public string Condicion { get; set; }
         public string SfFull { get; set; }
